Change design

+++ Code/MoneySaver/app/src/main/java/com/moneysaver/MainActivity.java
@@ -66,7 +66,7 @@ public class MainActivity extends AppCompatActivity
 
         TextView textView = findViewById(R.id.balance);
         textView.setTextSize(TypedValue.COMPLEX_UNIT_SP, 30);
-        textView.setText("Баланс: ".concat(Integer.toString(SQLite.getBalance(getBaseContext()))));
+        textView.setText("Баланс: ".concat(Double.toString(SQLite.getBalance(getBaseContext()))));
     }
 
     @Override
